Feat: 공통 타입 파일 생성 후 적용
